raise 404 when unknown url
consistent with other endpoints.

also refactor of notification_schema to separate some fns to a different file
